Move fixtures around to ensure proper location
